requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.3
django-celery-beat==2.1.0
django-celery-results==1.2.1
django-cors-headers==3.5.0
django-redis==4.12.1
django-timezone-field==4.0
djangorestframework==3.12.2
kombu==4.6.11
prompt-toolkit==3.0.8
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==1.3.0
wcwidth==0.2.5
